requirements.txt
python-telegram-bot>=21.0,<22.0
psycopg[binary]>=3.1.0,<4.0
imagehash>=4.3.0,<5.0
pillow>=10.0.0,<11.0
fuzzywuzzy>=0.18.0,<1.0
python-Levenshtein>=0.25.0,<1.0
requests>=2.31.0,<3.0
python-dotenv>=1.0.0,<2.0
confusable_homoglyphs>=3.2.0,<4.0
